+++ CHANGES.rst
@@ -4,6 +4,21 @@ Changelog
 master
 ------
 
+
+18.1.10
+-------
+
+* Display retype errors when stub application fails. Merge of #52, fixes #49.
+
+* Add ``--sample-count`` option to show the number of traces a given stub is
+  based on. Merge of #50, fixes #7. Thanks Tai-Lin.
+
+* Add ``monkeytype run -m`` for running a module as a script. Merge of
+  #41. Thanks Simon Gomizelj.
+
+* Add support for Django's ``cached_property`` decorator. Merge of #46, fixes
+  #9. Thanks Christopher J Wang.
+
 * Catch and log serialization exceptions instead of crashing. Fixes #38, merge
   of #39.